+++ test/dao/network-fee-withdraw.ts
@@ -3,6 +3,8 @@ import * as helpers from '../helpers/contract-helpers';
 import * as utils from '../helpers/utils';
 import { expect } from 'chai';
 import { GasGroup } from '../helpers/gas-usage';
+import { time } from "@nomicfoundation/hardhat-network-helpers";
+
 
 // Declare globals
 let ssvNetworkContract: any, minDepositAmount: any, burnPerBlock: any, networkFee: any;
@@ -26,13 +28,15 @@ describe('DAO Network Fee Withdraw Tests', () => {
 
     // Set network fee
     await ssvNetworkContract.updateNetworkFee(networkFee);
+    await utils.progressTime(172800); // 2 days
+    await ssvNetworkContract.updateNetworkFee(networkFee);
 
     // Register validators
     // cold register
     await helpers.DB.ssvToken.connect(helpers.DB.owners[6]).approve(helpers.DB.ssvNetwork.contract.address, '1000000000000000');
     await ssvNetworkContract.connect(helpers.DB.owners[6]).registerValidator(
       '0x221111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111119',
-      [1,2,3,4],
+      [1, 2, 3, 4],
       helpers.DataGenerator.shares(4),
       '1000000000000000',
       {
@@ -54,9 +58,30 @@ describe('DAO Network Fee Withdraw Tests', () => {
   });
 
   it('Withdraw network earnings emits "NetworkEarningsWithdrawn"', async () => {
+    const timestamp = await time.latest() + 1;
+    const releaseDate = timestamp + (86400 * 2);
+    const selector = ssvNetworkContract.interface.getSighash("withdrawNetworkEarnings(uint256)");
+
     const amount = await ssvNetworkContract.getNetworkEarnings();
     await expect(ssvNetworkContract.withdrawNetworkEarnings(amount
+    )).to.emit(ssvNetworkContract, 'FunctionLocked').withArgs(selector, releaseDate, helpers.DB.owners[0].address);
+    await utils.progressTime(172800); // 2 days
+
+    await expect(ssvNetworkContract.withdrawNetworkEarnings(amount
     )).to.emit(ssvNetworkContract, 'NetworkEarningsWithdrawn').withArgs(amount, helpers.DB.owners[0].address);
+  });
+
+  it('Withdraw network earnings before 2 days period reverts "FunctionIsLocked"', async () => {
+    const timestamp = await time.latest() + 1;
+    const releaseDate = timestamp + (86400 * 2);
+    const signature = ssvNetworkContract.interface.getSighash("withdrawNetworkEarnings(uint256)");
+
+    const amount = await ssvNetworkContract.getNetworkEarnings();
+    await expect(ssvNetworkContract.withdrawNetworkEarnings(amount)).to.emit(ssvNetworkContract, 'FunctionLocked').withArgs(signature, releaseDate, helpers.DB.owners[0].address);
+
+    await utils.progressTime(86400); // 1 day
+
+    await expect(ssvNetworkContract.withdrawNetworkEarnings(amount)).to.be.revertedWithCustomError(ssvNetworkContract, 'FunctionIsLocked');
   });
 
   it('Get withdrawable network earnings', async () => {
@@ -69,8 +94,10 @@ describe('DAO Network Fee Withdraw Tests', () => {
 
   it('Withdraw network earnings with not enough balance reverts "InsufficientBalance"', async () => {
     const amount = await ssvNetworkContract.getNetworkEarnings() * 2;
+    await ssvNetworkContract.withdrawNetworkEarnings(amount);
+    await utils.progressTime(172800); // 2 days
     await expect(ssvNetworkContract.withdrawNetworkEarnings(amount
-    )).to.be.revertedWithCustomError(ssvNetworkContract,'InsufficientBalance');
+    )).to.be.revertedWithCustomError(ssvNetworkContract, 'InsufficientBalance');
   });
 
   it('Withdraw network earnings from an address thats not the DAO reverts "caller is not the owner"', async () => {
